moved all files out, fixes for mypy, pylint, and sphinx
