+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_table_service_client_async.py
@@ -15,18 +15,19 @@ from azure.core.exceptions import HttpResponseError
 from azure.core.pipeline import AsyncPipeline
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
-from azure.data.tables import VERSION, LocationMode
-from azure.data.tables._generated.aio._azure_table_async import AzureTable
-from azure.data.tables._generated.models import TableServiceProperties, TableProperties, QueryOptions
-from azure.data.tables._models import service_stats_deserialize, service_properties_deserialize
-from ._base_client_async import AsyncStorageAccountHostsMixin, AsyncTransportWrapper
-from azure.data.tables.aio._policies_async import ExponentialRetry
-from azure.data.tables._deserialize import process_table_error
-from azure.data.tables.aio._table_client_async import TableClient
-from ._models import TablePropertiesPaged
-from .._shared._error import _validate_table_name
+
+from .. import VERSION, LocationMode
+from .._generated.aio._azure_table_async import AzureTable
+from .._generated.models import TableServiceProperties, TableProperties, QueryOptions
+from .._models import service_stats_deserialize, service_properties_deserialize
+from .._deserialize import process_table_error
+from .._error import _validate_table_name
 from .._table_service_client_base import TableServiceClientBase
 from .._models import Table
+from ._policies_async import ExponentialRetry
+from ._table_client_async import TableClient
+from ._base_client_async import AsyncStorageAccountHostsMixin, AsyncTransportWrapper
+from ._models import TablePropertiesPaged
 
 
 class TableServiceClient(AsyncStorageAccountHostsMixin, TableServiceClientBase):
