+++ sdk/tables/azure-data-tables/azure/data/tables/table_shared_access_signature.py
@@ -7,8 +7,8 @@ from typing import Union
 
 from ._models import AccountSasPermissions
 from ._models import TableServices
-from ._shared._common_conversion import _sign_string
-from ._shared._error import _validate_not_none
+from ._common_conversion import _sign_string
+from ._error import _validate_not_none
 from ._constants import X_MS_VERSION
 from .shared_access_signature import _SharedAccessHelper,\
     SharedAccessSignature, QueryStringConstants
